fix: change route grouping change quotes logic

+++ routes/web.php
@@ -29,24 +29,22 @@ Route::middleware(['guest'])->group(function () {
 Route::get('/logout', [LoginController::class, 'logout'])->name('logout')->middleware('auth');
 Route::get('set-language/{language}', [LanguageController::class, 'setLanguage'])->name('set-language');
 
-Route::group(['prefix' => 'admin/movies', 'middleware' => ['auth']], function () {
+Route::group(['middleware' => ['auth']], function () {
 	Route::controller(AdminMovieController::class)->group(function () {
-		Route::get('/', 'index')->name('adminpanel');
-		Route::post('/', 'store')->name('movies.store');
-		Route::patch('/{movie}', 'update')->name('movies.update');
-		Route::delete('/{movie}', 'destroy')->name('movies.delete');
-		Route::get('/{movie}/edit', 'edit')->name('movies.edit');
-		Route::get('/create', 'create')->name('movies.create');
+		Route::get('admin/movies', 'index')->name('adminpanel');
+		Route::get('admin/movies/create', 'create')->name('movies.create');
+		Route::post('admin/movies', 'store')->name('movies.store');
+		Route::get('admin/movies/{movie}/edit', 'edit')->name('movies.edit');
+		Route::patch('admin/movies/{movie}', 'update')->name('movies.update');
+		Route::delete('admin/movies/{movie}', 'destroy')->name('movies.delete');
 	});
-});
 
-Route::group(['prefix' => 'admin/movies/{movie}/quotes', 'middleware' => ['auth']], function () {
 	Route::controller(AdminQuoteController::class)->group(function () {
-		Route::get('/', 'index')->name('adminquotes');
-		Route::get('/create', 'create')->name('quotes.create');
-		Route::post('/', 'store')->name('quotes.store');
-		Route::delete('/{quote}', 'destroy')->name('quotes.delete');
-		Route::get('/{quote}/edit', 'edit')->name('quotes.edit');
-		Route::patch('/{quote}', 'update')->name('quotes.update');
+		Route::get('admin/quotes/', 'index')->name('adminquotes');
+		Route::get('admin/quotes/create', 'create')->name('quotes.create');
+		Route::post('admin/quotes/', 'store')->name('quotes.store');
+		Route::get('admin/quotes/{quote}/edit', 'edit')->name('quotes.edit');
+		Route::patch('admin/quotes/{quote}', 'update')->name('quotes.update');
+		Route::delete('admin/quotes/{quote}', 'destroy')->name('quotes.delete');
 	});
 });
